requirements.txt
discord.py>=2.0
apscheduler>=3.10
aiosqlite>=0.19
sqlalchemy
python-dotenv
pytest>=7
pytest-asyncio>=0.23
